merge z Bartka kodem

+++ super_bomberman/Crate.cpp
@@ -2,7 +2,7 @@
 
 
 
-Crate::Crate(BombermanGame *bomberman_game, const double &board_x, const double &board_y) : BoardObject(bomberman_game, board_x, board_y, false, true)
+Crate::Crate(BombermanGame *bomberman_game, const int &board_x, const int &board_y) : BoardObject(bomberman_game, board_x, board_y, false, true)
 {
 
 }
