+++ super_bomberman/Crate.h
@@ -4,7 +4,7 @@
 class Crate : public BoardObject
 {
 public:
-	Crate(BombermanGame *bomberman_game, const double &board_x, const double &board_y);
+	Crate(BombermanGame *bomberman_game, const int &board_x, const int &board_y);
 	~Crate();
 };
